+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased] (Preview)
 
+### Changed
+- (**Theme**) Improve automatic theme background colors
+
 ## [20251230.01] (r2937) - 2025-12-30
 
 ### Added